Move qsort_sanity to helper function
Makes it possible to track the debug symbol more easily.

Also rename std to std::sort

+++ src/main.cpp
@@ -130,7 +130,7 @@ int main(int argc, char** argv)
 
   bool checked = false;
 
-  for (size_t i = 0; i < arguments.runs; i++)
+  for (int64_t i = 0; i < arguments.runs; i++)
   {
     std::copy(orig_data.begin(), orig_data.end(), to_sort);
     const boost::timer::cpu_times res =
